Move EventSourceCache to other file
