+++ src/WinRT.Runtime/Interop/EventSourceCache.cs
@@ -1,0 +1,149 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Threading;
+
+namespace WinRT.Interop
+{
+    internal sealed class EventSourceCache
+    {
+        private static readonly ReaderWriterLockSlim cachesLock = new();
+        private static readonly ConcurrentDictionary<IntPtr, EventSourceCache> caches = new();
+
+        private readonly ConcurrentDictionary<int, System.WeakReference<State>> states = new();
+        private IWeakReference target;
+
+        private EventSourceCache(IWeakReference target, int index, System.WeakReference<State> state)
+        {
+            this.target = target;
+
+            SetState(index, state);
+        }
+
+        private EventSourceCache Update(IWeakReference target, int index, System.WeakReference<State> state)
+        {
+            // If target no longer exists, destroy cache
+            lock (this)
+            {
+                using var resolved = this.target.Resolve(IID.IID_IUnknown);
+                if (resolved == null)
+                {
+                    this.target = target;
+                    states.Clear();
+                }
+            }
+            SetState(index, state);
+            return this;
+        }
+
+        private System.WeakReference<State> GetState(int index)
+        {
+            // If target no longer exists, destroy cache
+            lock (this)
+            {
+                using var resolved = this.target.Resolve(IID.IID_IUnknown);
+                if (resolved == null)
+                {
+                    return null;
+                }
+            }
+
+            if (states.TryGetValue(index, out var weakState))
+            {
+                return weakState;
+            }
+            return null;
+        }
+
+        private void SetState(int index, System.WeakReference<State> state)
+        {
+            states[index] = state;
+        }
+
+        public static void Create(IObjectReference obj, int index, System.WeakReference<State> state)
+        {
+            // If event source implements weak reference support, track event registrations so that
+            // unsubscribes will work across garbage collections.  Note that most static/factory classes
+            // do not implement IWeakReferenceSource, so static codegen caching approach is also used.
+            IWeakReference target = null;
+#if !NET
+            try
+            {
+                var weakRefSource = (IWeakReferenceSource)typeof(IWeakReferenceSource).GetHelperType().GetConstructor(new[] { typeof(IObjectReference) }).Invoke(new object[] { obj });
+                if (weakRefSource == null)
+                {
+                    return;
+                }
+                target = weakRefSource.GetWeakReference();
+            }
+            catch (Exception)
+            {
+                return;
+            }
+#else
+            int hr = obj.TryAs<IUnknownVftbl>(IID.IID_IWeakReferenceSource, out var weakRefSource);
+            if (hr != 0)
+            {
+                return;
+            }
+
+            target = ABI.WinRT.Interop.IWeakReferenceSourceMethods.GetWeakReference(weakRefSource);
+#endif
+
+            cachesLock.EnterReadLock();
+            try
+            {
+                caches.AddOrUpdate(obj.ThisPtr,
+                    (IntPtr ThisPtr) => new EventSourceCache(target, index, state),
+                    (IntPtr ThisPtr, EventSourceCache cache) => cache.Update(target, index, state));
+            }
+            finally
+            {
+                cachesLock.ExitReadLock();
+            }
+        }
+
+        public static System.WeakReference<State> GetState(IObjectReference obj, int index)
+        {
+            if (caches.TryGetValue(obj.ThisPtr, out var cache))
+            {
+                return cache.GetState(index);
+            }
+
+            return null;
+        }
+
+        public static void Remove(IntPtr thisPtr, int index, System.WeakReference<State> state)
+        {
+            if (caches.TryGetValue(thisPtr, out var cache))
+            {
+#if !NET
+                // https://devblogs.microsoft.com/pfxteam/little-known-gems-atomic-conditional-removals-from-concurrentdictionary/
+                ((ICollection<KeyValuePair<int, System.WeakReference<State>>>)cache.states).Remove(
+                    new KeyValuePair<int, System.WeakReference<State>>(index, state));
+#else
+                cache.states.TryRemove(new KeyValuePair<int, System.WeakReference<State>>(index, state));
+#endif
+                // using double-checked lock idiom
+                if (cache.states.IsEmpty)
+                {
+                    cachesLock.EnterWriteLock();
+                    try
+                    {
+                        if (cache.states.IsEmpty)
+                        {
+                            caches.TryRemove(thisPtr, out var _);
+                        }
+                    }
+                    finally
+                    {
+                        cachesLock.ExitWriteLock();
+                    }
+                }
+            }
+        }
+    }
+}
